// frontend/web/src/services/API/proserve-wb-packaging-api/models/RecipeVersionEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RecipeVersionEntry
 */
export interface RecipeVersionEntry {
    /**
     * Recipe ID
     * @type {string}
     * @memberof RecipeVersionEntry
     */
    recipeId: string;
    /**
     * Recipe Name
     * @type {string}
     * @memberof RecipeVersionEntry
     */
    recipeName: string;
    /**
     * Recipe Version ID
     * @type {string}
     * @memberof RecipeVersionEntry
     */
    recipeVersionId: string;
    /**
     * Recipe Version Name
     * @type {string}
     * @memberof RecipeVersionEntry
     */
    recipeVersionName: string;
}

/**
 * Check if a given object implements the RecipeVersionEntry interface.
 */
export function instanceOfRecipeVersionEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipeId" in value;
    isInstance = isInstance && "recipeName" in value;
    isInstance = isInstance && "recipeVersionId" in value;
    isInstance = isInstance && "recipeVersionName" in value;

    return isInstance;
}

export function RecipeVersionEntryFromJSON(json: any): RecipeVersionEntry {
    return RecipeVersionEntryFromJSONTyped(json, false);
}

export function RecipeVersionEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecipeVersionEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipeId': json['recipeId'],
        'recipeName': json['recipeName'],
        'recipeVersionId': json['recipeVersionId'],
        'recipeVersionName': json['recipeVersionName'],
    };
}

export function RecipeVersionEntryToJSON(value?: RecipeVersionEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipeId': value.recipeId,
        'recipeName': value.recipeName,
        'recipeVersionId': value.recipeVersionId,
        'recipeVersionName': value.recipeVersionName,
    };
}